fix: migrate tests from com.vrpirates to com.vrhub package
- Migrated ConfigurationViewModelTest, StatusMappingTest, AnimationStateMachineTest
- Deleted orphaned test files that referenced old package
- Tests now use correct VRHub imports and compile successfully
