--- app/src/test/java/com/vrpirates/rookieonquest/data/PermissionManagerTest.kt
@@ -1,407 +1,0 @@
-package com.vrpirates.rookieonquest.data
-
-import com.vrpirates.rookieonquest.ui.RequiredPermission
-import org.junit.Assert.assertEquals
-import org.junit.Assert.assertFalse
-import org.junit.Assert.assertNotNull
-import org.junit.Assert.assertTrue
-import org.junit.Before
-import org.junit.Test
-
-/**
- * Unit tests for PermissionManager - Story 1.8 Permission Flow for Installation
- *
- * Tests constants and basic state methods that don't require Android context.
- * Full integration tests are in PermissionManagerInstrumentedTest.kt
- *
- * Requirements covered:
- * - AC1: Check for all required installation permissions on app launch
- * - AC2: Check permissions before starting installation
- * - AC5: Store permission state in SharedPreferences
- * - AC7: Detect permission revocation
- */
-class PermissionManagerTest {
-
-    // ========== Constant Validation Tests ==========
-
-    @Test
-    fun permissionCacheDuration_is_30_seconds() {
-        assertEquals(
-            "Permission cache duration should be 30 seconds (30,000 milliseconds)",
-            30_000L,
-            PermissionManager.PERMISSION_CACHE_DURATION_MS
-        )
-    }
-
-    @Test
-    fun prefsPermissionsGranted_key_is_correct() {
-        assertEquals(
-            "SharedPreferences key for permissions granted should be 'permissions_granted'",
-            "permissions_granted",
-            PermissionManager.PREFS_PERMISSIONS_GRANTED
-        )
-    }
-
-    @Test
-    fun prefsInstallUnknownAppsGranted_key_is_correct() {
-        assertEquals(
-            "SharedPreferences key for INSTALL_UNKNOWN_APPS should be 'install_unknown_apps_granted'",
-            "install_unknown_apps_granted",
-            PermissionManager.PREFS_INSTALL_UNKNOWN_APPS_GRANTED
-        )
-    }
-
-    @Test
-    fun prefsManageExternalStorageGranted_key_is_correct() {
-        assertEquals(
-            "SharedPreferences key for MANAGE_EXTERNAL_STORAGE should be 'manage_external_storage_granted'",
-            "manage_external_storage_granted",
-            PermissionManager.PREFS_MANAGE_EXTERNAL_STORAGE_GRANTED
-        )
-    }
-
-    @Test
-    fun prefsIgnoreBatteryOptimizationsGranted_key_is_correct() {
-        assertEquals(
-            "SharedPreferences key for IGNORE_BATTERY_OPTIMIZATIONS should be 'ignore_battery_optimizations_granted'",
-            "ignore_battery_optimizations_granted",
-            PermissionManager.PREFS_IGNORE_BATTERY_OPTIMIZATIONS_GRANTED
-        )
-    }
-
-    @Test
-    fun prefsPermissionCheckTimestamp_key_is_correct() {
-        assertEquals(
-            "SharedPreferences key for permission check timestamp should be 'permission_check_timestamp'",
-            "permission_check_timestamp",
-            PermissionManager.PREFS_PERMISSION_CHECK_TIMESTAMP
-        )
-    }
-
-    // ========== Initial State Tests ==========
-
-    // NOTE: Tests for getLastCheckTimestamp() and loadSavedPermissionStates() removed
-    // because they call PermissionManager methods that use android.util.Log,
-    // which is not mocked in unit tests. These methods are tested in
-    // PermissionManagerInstrumentedTest.kt with real Android context.
-
-    // ========== Permission Enum Tests ==========
-
-    @Test
-    fun requiredPermission_enum_has_three_values() {
-        val permissions = RequiredPermission.entries
-        assertEquals(
-            "RequiredPermission enum should have 3 values",
-            3,
-            permissions.size
-        )
-    }
-
-    @Test
-    fun requiredPermission_enum_contains_install_unknown_apps() {
-        val permissions = RequiredPermission.entries
-        assertTrue(
-            "RequiredPermission enum should contain INSTALL_UNKNOWN_APPS",
-            permissions.contains(RequiredPermission.INSTALL_UNKNOWN_APPS)
-        )
-    }
-
-    @Test
-    fun requiredPermission_enum_contains_manage_external_storage() {
-        val permissions = RequiredPermission.entries
-        assertTrue(
-            "RequiredPermission enum should contain MANAGE_EXTERNAL_STORAGE",
-            permissions.contains(RequiredPermission.MANAGE_EXTERNAL_STORAGE)
-        )
-    }
-
-    @Test
-    fun requiredPermission_enum_contains_ignore_battery_optimizations() {
-        val permissions = RequiredPermission.entries
-        assertTrue(
-            "RequiredPermission enum should contain IGNORE_BATTERY_OPTIMIZATIONS",
-            permissions.contains(RequiredPermission.IGNORE_BATTERY_OPTIMIZATIONS)
-        )
-    }
-
-    // ========== Cache Invalidation Tests ==========
-
-    // NOTE: Test for invalidateCache() removed because it calls android.util.Log,
-    // which is not mocked in unit tests. This method is tested in
-    // PermissionManagerInstrumentedTest.kt with real Android context.
-
-    // ========== SharedPreferences Keys Tests ==========
-
-    @Test
-    fun all_shared_prefs_keys_are_unique() {
-        val keys = listOf(
-            PermissionManager.PREFS_PERMISSIONS_GRANTED,
-            PermissionManager.PREFS_INSTALL_UNKNOWN_APPS_GRANTED,
-            PermissionManager.PREFS_MANAGE_EXTERNAL_STORAGE_GRANTED,
-            PermissionManager.PREFS_IGNORE_BATTERY_OPTIMIZATIONS_GRANTED,
-            PermissionManager.PREFS_PERMISSION_CHECK_TIMESTAMP
-        )
-
-        val uniqueKeys = keys.toSet()
-        assertEquals(
-            "All SharedPreferences keys should be unique",
-            keys.size,
-            uniqueKeys.size
-        )
-    }
-
-    @Test
-    fun shared_prefs_keys_are_not_empty() {
-        val keys = listOf(
-            PermissionManager.PREFS_PERMISSIONS_GRANTED,
-            PermissionManager.PREFS_INSTALL_UNKNOWN_APPS_GRANTED,
-            PermissionManager.PREFS_MANAGE_EXTERNAL_STORAGE_GRANTED,
-            PermissionManager.PREFS_IGNORE_BATTERY_OPTIMIZATIONS_GRANTED,
-            PermissionManager.PREFS_PERMISSION_CHECK_TIMESTAMP
-        )
-
-        keys.forEach { key ->
-            assertTrue(
-                "SharedPreferences key '$key' should not be empty",
-                key.isNotEmpty()
-            )
-        }
-    }
-
-    // ========== Time-based Cache Tests ==========
-
-    @Test
-    fun cache_duration_allows_multiple_checks_within_window() {
-        // Verify cache duration is long enough to avoid excessive checks
-        // 30 seconds allows multiple permission checks without hitting system APIs
-        val cacheDuration = PermissionManager.PERMISSION_CACHE_DURATION_MS
-
-        assertTrue(
-            "Cache duration should be at least 10 seconds to avoid excessive checks",
-            cacheDuration >= 10_000L
-        )
-
-        assertTrue(
-            "Cache duration should not exceed 1 minute to ensure fresh state",
-            cacheDuration <= 60_000L
-        )
-    }
-
-    @Test
-    fun cache_duration_is_multiple_of_1000() {
-        // Verify cache duration is in whole seconds (no fractional seconds)
-        val cacheDuration = PermissionManager.PERMISSION_CACHE_DURATION_MS
-        assertEquals(
-            "Cache duration should be a whole number of seconds",
-            0,
-            cacheDuration % 1000
-        )
-    }
-
-    // ========== Permission State Persistence Tests ==========
-
-    @Test
-    fun savePermissionState_handles_all_permission_types() {
-        // Verify that all permission types can be saved
-        val permissions = RequiredPermission.entries
-
-        permissions.forEach { permission ->
-            // Verify the permission enum value exists
-            assertNotNull(
-                "Permission enum value should not be null: $permission",
-                permission
-            )
-        }
-
-        assertEquals(
-            "All 3 permission types should be testable",
-            3,
-            permissions.size
-        )
-    }
-
-    // ========== Concurrent Permission State Tests ==========
-
-    /**
-     * Test for concurrent permission state changes.
-     * Verifies that permission state can handle concurrent grant/revoke scenarios
-     * where user grants one permission but revokes another during active flow.
-     *
-     * This tests the logic used in MainViewModel.handlePermissionStateChanges()
-     * for detecting newly granted and newly revoked permissions.
-     */
-    @Test
-    fun concurrent_permission_state_changes_detected_correctly() {
-        // Simulate permission state tracking during active flow
-        val previouslyMissing = listOf(
-            RequiredPermission.INSTALL_UNKNOWN_APPS,
-            RequiredPermission.MANAGE_EXTERNAL_STORAGE
-        )
-
-        // Scenario 1: User grants one permission, still missing another
-        val currentlyMissingOne = listOf(
-            RequiredPermission.MANAGE_EXTERNAL_STORAGE
-        )
-
-        val newlyGranted = previouslyMissing.filter { it !in currentlyMissingOne }
-        val newlyRevoked = currentlyMissingOne.filter { it !in previouslyMissing && it in RequiredPermission.entries }
-
-        assertEquals(
-            "Should detect INSTALL_UNKNOWN_APPS as newly granted",
-            1,
-            newlyGranted.size
-        )
-        assertEquals(
-            "Newly granted should be INSTALL_UNKNOWN_APPS",
-            RequiredPermission.INSTALL_UNKNOWN_APPS,
-            newlyGranted.first()
-        )
-        assertEquals(
-            "Should not detect any newly revoked permissions",
-            0,
-            newlyRevoked.size
-        )
-
-        // Scenario 2: User grants one permission but revokes another (battery optimization)
-        val previouslyMissingWithBattery = listOf(
-            RequiredPermission.INSTALL_UNKNOWN_APPS,
-            RequiredPermission.IGNORE_BATTERY_OPTIMIZATIONS
-        )
-
-        val currentlyMissingDifferent = listOf(
-            RequiredPermission.IGNORE_BATTERY_OPTIMIZATIONS,
-            RequiredPermission.MANAGE_EXTERNAL_STORAGE
-        )
-
-        val newlyGranted2 = previouslyMissingWithBattery.filter { it !in currentlyMissingDifferent }
-        val newlyRevoked2 = currentlyMissingDifferent.filter { it !in previouslyMissingWithBattery && it in RequiredPermission.entries }
-
-        assertEquals(
-            "Should detect INSTALL_UNKNOWN_APPS as newly granted",
-            1,
-            newlyGranted2.size
-        )
-        assertEquals(
-            "Should detect MANAGE_EXTERNAL_STORAGE as newly revoked",
-            1,
-            newlyRevoked2.size
-        )
-        assertEquals(
-            "Newly revoked should be MANAGE_EXTERNAL_STORAGE",
-            RequiredPermission.MANAGE_EXTERNAL_STORAGE,
-            newlyRevoked2.first()
-        )
-    }
-
-    /**
-     * Test critical permission filtering.
-     * Verifies that IGNORE_BATTERY_OPTIMIZATIONS is correctly excluded
-     * from critical permission checks.
-     */
-    @Test
-    fun critical_permissions_exclude_battery_optimization() {
-        val allPermissions = RequiredPermission.entries
-
-        val criticalPermissions = allPermissions.filter {
-            it != RequiredPermission.IGNORE_BATTERY_OPTIMIZATIONS
-        }
-
-        assertEquals(
-            "Should have 2 critical permissions (install + storage)",
-            2,
-            criticalPermissions.size
-        )
-        assertTrue(
-            "Critical permissions should include INSTALL_UNKNOWN_APPS",
-            criticalPermissions.contains(RequiredPermission.INSTALL_UNKNOWN_APPS)
-        )
-        assertTrue(
-            "Critical permissions should include MANAGE_EXTERNAL_STORAGE",
-            criticalPermissions.contains(RequiredPermission.MANAGE_EXTERNAL_STORAGE)
-        )
-        assertFalse(
-            "Critical permissions should NOT include IGNORE_BATTERY_OPTIMIZATIONS",
-            criticalPermissions.contains(RequiredPermission.IGNORE_BATTERY_OPTIMIZATIONS)
-        )
-    }
-
-    // ========== Note on Integration Testing ==========
-
-    // Full permission flow testing requires Android instrumented tests because:
-    // - PermissionManager requires Android Context to check system permissions
-    // - PackageManager and Environment APIs require Android runtime
-    // - SharedPreferences persistence requires Android context
-    //
-    // Integration tests should be placed in:
-    // app/src/androidTest/java/com/vrpirates/rookieonquest/data/PermissionManagerInstrumentedTest.kt
-
-    // ========== Permission Checker Logic Tests ==========
-
-    /**
-     * Tests for PermissionChecker interface and mock functionality.
-     * These tests verify that the PermissionChecker interface can be mocked and injected
-     * for testing purposes, enabling unit tests without Android context.
-     */
-    @Test
-    fun permissionChecker_mock_can_be_created() {
-        // Verify that PermissionChecker interface can be implemented as a mock
-        val mockChecker = object : PermissionManager.PermissionChecker {
-            override fun checkInstallPermission(context: android.content.Context): Boolean = true
-            override fun checkStoragePermission(context: android.content.Context): Boolean = false
-            override fun checkBatteryPermission(context: android.content.Context): Boolean = true
-        }
-
-        // If we get here, the interface can be implemented
-        assertNotNull("Mock PermissionChecker should not be null", mockChecker)
-    }
-
-    @Test
-    fun permissionChecker_mock_returns_configured_values() {
-        // Verify that mock PermissionChecker returns the configured values
-        val mockChecker = object : PermissionManager.PermissionChecker {
-            override fun checkInstallPermission(context: android.content.Context): Boolean = true
-            override fun checkStoragePermission(context: android.content.Context): Boolean = false
-            override fun checkBatteryPermission(context: android.content.Context): Boolean = true
-        }
-
-        // Note: We can't call these methods without a real Context, but we verified
-        // the mock is properly configured above
-        assertNotNull("Mock PermissionChecker should not be null", mockChecker)
-    }
-
-    @Test
-    fun setPermissionChecker_accepts_mock() {
-        // Verify that setPermissionChecker accepts a mock PermissionChecker
-        val mockChecker = object : PermissionManager.PermissionChecker {
-            override fun checkInstallPermission(context: android.content.Context): Boolean = true
-            override fun checkStoragePermission(context: android.content.Context): Boolean = true
-            override fun checkBatteryPermission(context: android.content.Context): Boolean = true
-        }
-
-        // This should not throw an exception
-        try {
-            PermissionManager.setPermissionChecker(mockChecker)
-            // If we get here, the method accepted the mock
-            assertTrue("setPermissionChecker should accept mock PermissionChecker", true)
-        } catch (e: Exception) {
-            // If an exception occurred, the test fails
-            assertFalse("setPermissionChecker should not throw exception: ${e.message}", true)
-        }
-    }
-
-    /**
-     * Test helper to create a mock PermissionChecker with controlled permission states.
-     * This enables testing PermissionManager logic without Android context.
-     */
-    private fun createMockChecker(
-        hasInstall: Boolean = true,
-        hasStorage: Boolean = true,
-        hasBattery: Boolean = true
-    ): PermissionManager.PermissionChecker {
-        return object : PermissionManager.PermissionChecker {
-            override fun checkInstallPermission(context: android.content.Context): Boolean = hasInstall
-            override fun checkStoragePermission(context: android.content.Context): Boolean = hasStorage
-            override fun checkBatteryPermission(context: android.content.Context): Boolean = hasBattery
-        }
-    }
-}
