--- app/src/test/java/com/vrpirates/rookieonquest/data/WakeLockManagerTest.kt
@@ -1,103 +1,0 @@
-package com.vrpirates.rookieonquest.data
-
-import org.junit.Assert.assertEquals
-import org.junit.Assert.assertFalse
-import org.junit.Test
-
-/**
- * Unit tests for WakeLockManager - Story 1.6 Wake Lock Implementation
- *
- * Tests constants and basic state methods that don't require Android context.
- * Full integration tests are in WakeLockManagerInstrumentedTest.kt
- *
- * Requirements covered:
- * - AC5: CPU wake lock prevents Quest sleep (NFR-P11, FR55)
- * - Timeout safety (max 30 minutes)
- */
-class WakeLockManagerTest {
-
-    // ========== Constant Validation Tests ==========
-
-    @Test
-    fun maxWakeLockTimeoutMs_is_30_minutes() {
-        val thirtyMinutesInMs = 30L * 60L * 1000L
-        assertEquals(
-            "Max wake lock timeout should be 30 minutes in milliseconds",
-            thirtyMinutesInMs,
-            WakeLockManager.MAX_WAKE_LOCK_TIMEOUT_MS
-        )
-    }
-
-    @Test
-    fun maxWakeLockTimeoutMs_equals_1800000_ms() {
-        assertEquals(
-            "Max wake lock timeout should be exactly 1,800,000 milliseconds",
-            1_800_000L,
-            WakeLockManager.MAX_WAKE_LOCK_TIMEOUT_MS
-        )
-    }
-
-    // ========== Initial State Tests ==========
-
-    @Test
-    fun isHeld_returns_false_when_never_acquired() {
-        // Note: This test assumes no previous test acquired the wake lock
-        // In a clean state, isHeld should return false
-        assertFalse(
-            "Wake lock should not be held initially",
-            WakeLockManager.isHeld()
-        )
-    }
-
-    @Test
-    fun getHeldDurationMs_returns_0_when_not_held() {
-        // When wake lock is not held, duration should be 0
-        val duration = WakeLockManager.getHeldDurationMs()
-        assertEquals(
-            "Duration should be 0 when wake lock is not held",
-            0L,
-            duration
-        )
-    }
-
-    @Test
-    fun hasExceededTwoMinutes_returns_false_when_not_held() {
-        // When wake lock is not held, hasExceededTwoMinutes should return false
-        assertFalse(
-            "hasExceededTwoMinutes should return false when wake lock is not held",
-            WakeLockManager.hasExceededTwoMinutes()
-        )
-    }
-
-    // ========== Release Safety Tests ==========
-    // NOTE: Tests that call release() or forceRelease() are in WakeLockManagerInstrumentedTest
-    // because they trigger Android Log calls which require mocking or instrumented tests.
-
-    // ========== Reference Counting Tests ==========
-    // NOTE: Reference counting behavior tests are in WakeLockManagerInstrumentedTest
-    // because they require calling acquire/release methods that use Android Log.
-
-    // ========== Two Minute Threshold Tests ==========
-
-    @Test
-    fun twoMinuteThreshold_is_120000_ms() {
-        // Verify the 2-minute threshold calculation used in hasExceededTwoMinutes()
-        val twoMinutesInMs = 2L * 60L * 1000L
-        assertEquals(
-            "Two minutes should be 120,000 milliseconds",
-            120_000L,
-            twoMinutesInMs
-        )
-    }
-
-    // ========== Extraction Progress Throttle Tests ==========
-
-    @Test
-    fun extractionProgressThrottleMs_is_1_second() {
-        assertEquals(
-            "Extraction progress throttle should be 1000ms (1Hz minimum update rate)",
-            1000L,
-            Constants.EXTRACTION_PROGRESS_THROTTLE_MS
-        )
-    }
-}
