--- app/src/test/java/com/vrpirates/rookieonquest/network/SecureUpdateFlowTest.kt
@@ -1,201 +1,0 @@
-package com.vrpirates.rookieonquest.network
-
-import kotlinx.coroutines.ExperimentalCoroutinesApi
-import kotlinx.coroutines.delay
-import kotlinx.coroutines.test.runTest
-import org.junit.Assert.*
-import org.junit.Test
-import java.io.IOException
-
-/**
- * Logical integration tests for the Secure Update flow.
- * 
- * These tests verify the core business logic of the update flow:
- * 1. Version comparison (SemVer-like with pre-release support)
- * 2. Retry logic with exponential backoff
- * 3. Resumable download logic (Range header and append mode)
- * 
- * These tests use the same logic implemented in MainViewModel.kt to ensure correctness.
- */
-@OptIn(ExperimentalCoroutinesApi::class)
-class SecureUpdateFlowTest {
-
-    // ========== Version Comparison Tests ==========
-
-    /**
-     * Re-implementation of MainViewModel.isVersionNewer for unit testing.
-     * Logic matches MainViewModel.kt lines 1428-1463.
-     */
-    private fun isVersionNewer(latest: String, current: String): Boolean {
-        // Simple SemVer-like comparison
-        val latestBase = latest.split('-')[0]
-        val currentBase = current.split('-')[0]
-        
-        val latestParts = latestBase.split('.').mapNotNull { it.filter { c -> c.isDigit() }.toIntOrNull() }
-        val currentParts = currentBase.split('.').mapNotNull { it.filter { c -> c.isDigit() }.toIntOrNull() }
-
-        val maxLength = maxOf(latestParts.size, currentParts.size)
-        for (i in 0 until maxLength) {
-            val latestPart = latestParts.getOrElse(i) { 0 }
-            val currentPart = currentParts.getOrElse(i) { 0 }
-            if (latestPart > currentPart) return true
-            if (latestPart < currentPart) return false
-        }
-        
-        // If version bases are identical, a version with a pre-release tag (hyphen)
-        // is considered OLDER than a version without one.
-        val latestHasPre = latest.contains('-')
-        val currentHasPre = current.contains('-')
-        
-        if (!latestHasPre && currentHasPre) return true // 2.5.0 is newer than 2.5.0-rc
-        if (latestHasPre && !currentHasPre) return false // 2.5.0-rc is older than 2.5.0
-        
-        // If both have pre-release tags, compare them alphabetically
-        if (latestHasPre && currentHasPre) {
-            val latestPre = latest.substringAfter('-')
-            val currentPre = current.substringAfter('-')
-            return latestPre > currentPre
-        }
-
-        return false
-    }
-
-    @Test
-    fun testVersionComparison() {
-        // Basic versions
-        assertTrue(isVersionNewer("2.5.0", "2.4.9"))
-        assertTrue(isVersionNewer("2.5.1", "2.5.0"))
-        assertTrue(isVersionNewer("3.0.0", "2.9.9"))
-        assertFalse(isVersionNewer("2.5.0", "2.5.0"))
-        assertFalse(isVersionNewer("2.4.0", "2.5.0"))
-
-        // Differing lengths
-        assertTrue(isVersionNewer("2.5", "2.4.1"))
-        assertFalse(isVersionNewer("2.5.0", "2.5")) // Equivalent versions
-        assertTrue(isVersionNewer("2.5.1", "2.5"))
-        
-        // Pre-release tags
-        assertTrue(isVersionNewer("2.5.0", "2.5.0-rc.1")) // Release is newer than pre-release
-        assertTrue(isVersionNewer("2.5.0-rc.2", "2.5.0-rc.1")) // Higher pre-release number
-        assertFalse(isVersionNewer("2.5.0-rc.1", "2.5.0")) // Pre-release is older than release
-        
-        // Prefixes (logic in ViewModel handles lowercase and 'v' removal before calling isVersionNewer)
-        assertTrue(isVersionNewer("2.5.0", "2.4.0"))
-    }
-
-    // ========== Retry Logic Tests ==========
-
-    @Test
-    fun testRetryLogicWithExponentialBackoff() = runTest {
-        var attempts = 0
-        var totalDelay = 0L
-        val maxRetries = 3
-        var currentDelay = 1000L
-
-        // Simulate checkForAppUpdates logic (MainViewModel.kt lines 1374-1412)
-        repeat(maxRetries) { attempt ->
-            attempts++
-            try {
-                // Simulate network failure for first 2 attempts
-                if (attempts <= 2) throw IOException("Network error")
-                
-                // Success on 3rd attempt
-                return@repeat
-            } catch (e: IOException) {
-                if (attempt < maxRetries - 1) {
-                    totalDelay += currentDelay
-                    // We don't actually delay in runTest, we just record it
-                    currentDelay *= 2
-                } else {
-                    fail("Should have succeeded on 3rd attempt")
-                }
-            }
-        }
-
-        assertEquals(3, attempts)
-        assertEquals(3000L, totalDelay) // 1000 + 2000
-    }
-
-    @Test
-    fun testRetryLogicFailureAfterMaxAttempts() = runTest {
-        var attempts = 0
-        val maxRetries = 3
-        var failed = false
-
-        repeat(maxRetries) { attempt ->
-            attempts++
-            try {
-                throw IOException("Permanent network error")
-            } catch (e: IOException) {
-                if (attempt == maxRetries - 1) {
-                    failed = true
-                }
-                // skip delay in test
-            }
-        }
-
-        assertEquals(3, attempts)
-        assertTrue(failed)
-    }
-
-    // ========== Resumable Download Logic Tests ==========
-
-    /**
-     * Test that when server doesn't support Range header (returns 200),
-     * the partial file is properly overwritten instead of appended to.
-     * This ensures correct behavior when transitioning from resume to full download.
-     */
-    @Test
-    fun testServerWithoutRangeHeaderSupport() {
-        // Simulate downloadAndInstallUpdate logic (MainViewModel.kt lines 1477-1502)
-
-        // Case: Partial file exists but server doesn't support Range header
-        val partialFileSize = 1024L // Partial file of 1KB exists
-
-        // When Range header is sent but server returns 200 (full content)
-        val responseCode = 200 // Server doesn't support Range, returns full file
-        val isResume = responseCode == 206
-
-        // Verify isResume is false when server returns 200
-        assertFalse("Server returning 200 should NOT be treated as resume", isResume)
-
-        // Verify append mode would be false (overwrite) when isResume is false
-        // This matches: FileOutputStream(targetFile, isResume) - false = overwrite
-        // Even with partialFileSize of 1KB, append mode is false because server returned 200
-        val appendMode = isResume
-        assertFalse("Partial file should be overwritten when server returns 200 (ignoring existing partial)", appendMode)
-
-        // Case: Server properly supports Range - returns 206
-        val responseCodeWithRange = 206
-        val isResumeWithRange = responseCodeWithRange == 206
-        // With partial file and 206 response, resume is possible
-        val canResume = partialFileSize > 0 && isResumeWithRange
-
-        assertTrue("Server returning 206 should be treated as resume", isResumeWithRange)
-        assertTrue("Partial file should be appended to when server returns 206 with existing partial", canResume)
-    }
-
-    @Test
-    fun testResumableDownloadParameters() {
-        // Simulate downloadAndInstallUpdate logic (MainViewModel.kt lines 1478-1488)
-        
-        // Case 1: No existing file
-        val initialDownloaded1 = 0L
-        val rangeHeader1 = if (initialDownloaded1 > 0) "bytes=$initialDownloaded1-" else null
-        assertNull(rangeHeader1)
-
-        // Case 2: Partial file exists
-        val initialDownloaded2 = 1024L
-        val rangeHeader2 = if (initialDownloaded2 > 0) "bytes=$initialDownloaded2-" else null
-        assertEquals("bytes=1024-", rangeHeader2)
-        
-        // Case 3: Response handling (MainViewModel.kt lines 1493-1494)
-        val responseCode = 206 // Partial Content
-        val isResume = responseCode == 206
-        assertTrue(isResume)
-        
-        val normalResponseCode = 200
-        val isNotResume = normalResponseCode == 206
-        assertFalse(isNotResume)
-    }
-}
